type casting bug

+++ helper/vga.c
@@ -335,7 +335,7 @@ void plotTimeDomain(point reference, int width, int height,
         amplitude /= sample_per_pixel;
         amplitude = amplitude < 0 ? 0 : amplitude;
 
-        int line_height = amplitude / max_sample_amplitude * (height - 2*axes_offset);
+        int line_height = (int)(((float)amplitude / (float)max_sample_amplitude) * (height - 2*axes_offset));
         drawLine((point){x, reference.y + (line_height/2)}, (point){x, reference.y - (line_height/2)}, LINE_COLOR, false);
     }
 }
